requirements.txt
JPype1==1.4.0
